requirements.txt
opencv-python-headless
paho-mqtt
requests
numpy
